+++ R/staticimports.R
@@ -89,11 +89,11 @@ str_extract <- function(string, pattern) {
 	)
 }
 
-str_extract_all <- function(string, pattern) {
+str_extract_all <- function(string, pattern, simplify = FALSE) {
 	is_fixed <- inherits(pattern, "fixed")
 	ignore.case <- isTRUE(attr(pattern, "options")$case_insensitive)
 
-	regmatches(
+	result <- regmatches(
 		x = string,
 		m = gregexpr(
 			pattern = pattern,
@@ -103,6 +103,20 @@ str_extract_all <- function(string, pattern) {
 			fixed = is_fixed
 		)
 	)
+
+	if (isTRUE(simplify)) {
+		max_length <- max(lengths(result))
+
+		result <- t(vapply(
+			result,
+			function(x) x[seq_len(max_length)],
+			character(max_length)
+		))
+
+		result[is.na(result)] <- ""
+	}
+
+	result
 }
 
 str_match <- function(string, pattern) {